# Add file system information to each event