--- a/libbeat/reader/readfile/metafields_test.go
+++ b/libbeat/reader/readfile/metafields_test.go
@@ -19,7 +19,9 @@ package readfile
 
 import (
 	"io"
+	"os"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
@@ -27,19 +29,19 @@ import (
 	"github.com/elastic/elastic-agent-libs/mapstr"
 )
 
-func TestMetaFieldsOffset(t *testing.T) {
+func TestMetaFields(t *testing.T) {
 	messages := []reader.Message{
-		reader.Message{
+		{
 			Content: []byte("my line"),
 			Bytes:   7,
 			Fields:  mapstr.M{},
 		},
-		reader.Message{
+		{
 			Content: []byte("my line again"),
 			Bytes:   13,
 			Fields:  mapstr.M{},
 		},
-		reader.Message{
+		{
 			Content: []byte(""),
 			Bytes:   10,
 			Fields:  mapstr.M{},
@@ -48,7 +50,8 @@ func TestMetaFieldsOffset(t *testing.T) {
 
 	path := "test/path"
 	offset := int64(0)
-	in := &FileMetaReader{msgReader(messages), path, offset}
+
+	in := &FileMetaReader{msgReader(messages), path, createTestFileInfo(), "hash", offset}
 	for {
 		msg, err := in.Next()
 		if err == io.EOF {
@@ -60,15 +63,18 @@ func TestMetaFieldsOffset(t *testing.T) {
 			expectedFields = mapstr.M{
 				"log": mapstr.M{
 					"file": mapstr.M{
-						"path": path,
+						"path":        path,
+						"fingerprint": "hash",
 					},
 					"offset": offset,
 				},
 			}
+			checkFields(t, expectedFields, msg.Fields)
+		} else {
+			require.Equal(t, expectedFields, msg.Fields)
 		}
 		offset += int64(msg.Bytes)
 
-		require.Equal(t, expectedFields, msg.Fields)
 		require.Equal(t, offset, in.offset)
 	}
 }
@@ -96,3 +102,17 @@ func (r *messageReader) Next() (reader.Message, error) {
 func (r *messageReader) Close() error {
 	return nil
 }
+
+type testFileInfo struct {
+	name string
+	size int64
+	time time.Time
+	sys  interface{}
+}
+
+func (t testFileInfo) Name() string       { return t.name }
+func (t testFileInfo) Size() int64        { return t.size }
+func (t testFileInfo) Mode() os.FileMode  { return 0 }
+func (t testFileInfo) ModTime() time.Time { return t.time }
+func (t testFileInfo) IsDir() bool        { return false }
+func (t testFileInfo) Sys() interface{}   { return t.sys }