--- a/libbeat/reader/readfile/metafields_windows_test.go
+++ b/libbeat/reader/readfile/metafields_windows_test.go
@@ -1,0 +1,69 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package readfile
+
+import (
+	"os"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/elastic/elastic-agent-libs/mapstr"
+)
+
+func createTestFileInfo() os.FileInfo {
+	return testFileInfo{
+		name: "filename",
+		size: 42,
+		time: time.Now(),
+		sys: struct {
+			idxhi uint
+			idxlo uint
+			vol   uint
+		}{
+			idxhi: 100,
+			idxlo: 200,
+			vol:   300,
+		},
+	}
+}
+
+func checkFields(t *testing.T, expected, actual mapstr.M) {
+	t.Helper()
+
+	idxlo, err := actual.GetValue(idxloKey)
+	require.NoError(t, err)
+	require.Equal(t, "100", idxlo)
+	err = actual.Delete(idxloKey)
+	require.NoError(t, err)
+
+	idxhi, err := actual.GetValue(idxhiKey)
+	require.NoError(t, err)
+	require.Equal(t, "200", idxhi)
+	err = actual.Delete(idxhiKey)
+	require.NoError(t, err)
+
+	vol, err := actual.GetValue(volKey)
+	require.NoError(t, err)
+	require.Equal(t, "300", vol)
+	err = actual.Delete(volKey)
+	require.NoError(t, err)
+
+	require.Equal(t, expected, actual)
+}