--- a/libbeat/reader/readfile/metafields_windows_test.go
+++ b/libbeat/reader/readfile/metafields_windows_test.go
@@ -35,7 +35,7 @@ type winTestInfo struct {
 }
 
 func createTestFileInfo() os.FileInfo {
-	return testFileInfo{
+	return winTestInfo{
 		testFileInfo: testFileInfo{
 			name: "filename",
 			size: 42,