@ycombinator 2023-07-14
> ```
> "log": {
>     "offset": 0,
>     "file": {
>       "inode": "119400347",
>       "fingerprint": "2edc986847e209b4016e141a6dc8716d3207350f416969382d431539bf292e4a",
>       "path": "/path/to/your/logs/some.log",
>       "device": "16777234"
>     }
>   }
> ```
> 

> ```
> "log": {
>     "offset": 0,
>     "file": {
>       "path": "C:\\Users\\Admin\\Desktop\\test\\fs_metadata\\logs\\some.log",
>       "idxhi": "327680",
>       "idxlo": "99308",
>       "vol": "2860917223",
>       "fingerprint": "0e9d2f73e97662c9a156111ea556fc536782043f2a1cd69fecbe3b69cab64c9d"
>     }
> ```
> 

Since we're adding what looks like 3-4 **string** fields to every event, I wonder how this will impact performance. Should we do some benchmarking first? Or put this behind a configuration setting (either enabled by default or disabled by default)?