@ycombinator 2023-07-14
Also, _should_ fields like `idxhi` and `idxlo` be string fields? I don't have any domain knowledge about these fields but they _seem_ like fields that might be useful for range queries, so integers might be more appropriate?