@rdner 2023-07-14
@ycombinator we've already [discussed](https://github.com/elastic/beats/issues/35801#issuecomment-1597266413) the performance impact in the original issue and agreed the value for us is more important.

The biggest field here is the fingerprint which is already an opt-in behaviour, this field is present only if the fingerprint mode is active. The rest are just very short strings.