@rdner 2023-07-14
@ycombinator also adding to my previous comment: the length of all the added fields is constant and they're attached only to events produced by the filestream input. So, the impact is not global.