@ycombinator 2023-07-14
It makes me a bit unesasy that we don't actually know what the performance impact of these additional fields on every filestream event would be.  But I see the points that the sizes of these fields are fixed and also that in terms of a % increase compared to the size of the overall event, the increase might not be significant.

The only question I have remaining, then, is https://github.com/elastic/beats/pull/36065#issuecomment-1635938924.