@ycombinator 2023-07-14
> This was requested by some customers and it makes it easier for us to troubleshoot.

Given that this was requested by _some_ customers, for _troubleshooting_ reasons, and also given that we don't actually _know_ the performance impact (we're speculating that it won't be significant, and I tend to agree with this speculation, but it's still not definitive proof), I'd really be inclined to put the inclusion of the fields behind a configuration setting that's disabled by default.  That way, when we/a specific customer needs this information for troubleshooting, they can enable the setting and when they're done troubleshooting, they can disable it (or leave it enabled if the performance impact is not material).