@rdner 2023-07-14
@ycombinator don't get me wrong, I'm all for testing the performance here. We can do it once @alexsapran is available. If you think we should hold this PR until then, I'm fine with that.

Regarding the use of integers in https://github.com/elastic/beats/pull/36065#issuecomment-1635938924 I made the change here https://github.com/elastic/beats/pull/36065/commits/abc9643e44385bfd4d74ae69a9c14b9a631b029c