@ycombinator 2023-07-14
> @ycombinator don't get me wrong, I'm all for testing the performance here. We can do it once @alexsapran is available. If you think we should hold this PR until then, I'm fine with that.

I'm 99% sure the performance testing is going to agree with our speculation (that the performance impact is non-existent or negligible).  But I think it's worth waiting for it only because once this is released we won't have a way to undo it in case there are negative consequences.  

That said, we have another month to `8.10.0` feature freeze so I'd be okay with merging this PR as-is provided we can be assured of doing the performance testing before that deadline.

> Regarding the use of integers in [#36065 (comment)](https://github.com/elastic/beats/pull/36065#issuecomment-1635938924) I made the change here [abc9643](https://github.com/elastic/beats/commit/abc9643e44385bfd4d74ae69a9c14b9a631b029c)

Thanks!