@rdner 2023-07-14
@pierrehilbert @nimarezainia @cmacknz opinions?

Should we add a new flag to the filestream input configuration to make these fields optional?

The maximum impact from this change is an event size increase by:

* 141 bytes on Windows (`"idxhi": 4294967295,"idxlo": 4294967295,"vol": 4294967295,"fingerprint": "2edc986847e209b4016e141a6dc8716d3207350f416969382d431539bf292e4a",`)
* 144 bytes on Unix (`"inode": 18446744073709551615,"device": 18446744073709551615,"fingerprint": "2edc986847e209b4016e141a6dc8716d3207350f416969382d431539bf292e4a",`)

I personally think that the input configuration is already too polluted with options, I'm not sure if it's worth to add a new option to save ~145 bytes.