@ycombinator 2023-07-14
To be clear, my suggestion for a new configuration flag to enable/disable these fields is only because we haven't (yet) performance-tested the impact of this change — a change that adds ~145 bytes to every filestream event.  If we do the performance testing and it proves that the impact is nonexistent or negligible, I don't think we need a flag.