@pierrehilbert 2023-07-16
As mentioned in the issue itself when you created it, I agree that we should validate the perf impact of this change to ensure that it's as little as what we imagine.
@alexsapran can you help us here or, better option, give us everything to be able to run the tests by ourself?