@rdner 2023-07-16
@pierrehilbert see my comment above https://github.com/elastic/beats/pull/36065#issuecomment-1636025777