@alexsapran 2023-07-17
Hi all, sorry for the late replay, I was out sick. 

Looking at the discussion I can definitely help, but at the same time, I want to make sure we are talking about the same items when we say performance. 
What performance implications do we anticipate this PR to have? Is only the additional 145bytes for each event on the output side or do we need to calculate the fingerprint and file info for each event?
I can sync with Denis and provide an environment in which we can test this out, but from reading this it's not clear to be which aspect of the performance impact we are talking about.