@rdner 2023-07-17
@ycombinator as @alexsapran pointed out to me, in ECS https://www.elastic.co/guide/en/ecs/current/ecs-file.html#field-file-device `device` is a keyword for some reason.

So, what should we do, keep the integer because it makes sense, or be ECS compliant with a string value?

I suppose we cannot make a change to ECS, can we?