@alexsapran 2023-07-17
Overall it would be nice to backport those changes back to ECS so we keep the FB output ECS compliant when possible, I know this is more involved, but it's something to keep an eye out for future work we do especially when adding new fields.