@rdner 2023-07-17
@alexsapran well technically, the changes in this PR are compliant. It's just ECS does not have additional fields that we have here. I'm not even sure it should.