@alexsapran 2023-07-17
I run the benchmarks for the use case of JSON formatted log and unstructured format.

### JSON
The data show that we were able to achieve the same amount of EPS. I noticed a slight increase in the output throughput metrics, from 20.9MB/s to 21.6MB/s, which is expected given that we are increasing the data we ship.

### Unstructured RAW
When I was running the benchmark I noticed that the EPS was increased after this PR, not including the changes from f8582d6 which was pushed after I started. 
Similar to the JSON we saw also an increase in the output throughput, from 30Mb/s to 33.1MB/s which is normal given the additional data we sent. 

All of the above was done with compression level 0, which is the default compression.

I am re-running the unstructured tests and looking into them. 

Overall I don't see this PR adding any performance regression on the output side of the process.